fix: add default key manager type check tests when HSM enabled
